Changed the name of the BroadCastReceiver to ProjectBroadcastReceiver to make room to add ServerBroadcastReceiver..
Also added db helper functions to the GlobalDatabaseHelper to create the servers table, get the list of servers to populate the servers dialog, and create a new server.

I also changed the name of the ArbiterDialogFragments registerCustomListeners method, because it's really just an entry point to do anything with the view before displaying the dialog.

I also made some minor formatting changes