Committing everything at once to avoid getting the app into a bad state…
Added a BroadcastReceiver, Loader, and Adapter for layers so that the layers dialog gets updated when new layers are added.  I also updated the layers_dialog to use a regular list view instead of the custom list view that was being used as a placeholder to demonstrate what the dialog would look like when layers get added.

Layers get added now, but I just noticed an error with the AddLayersDialog not updating properly on server selection change after adding a layer.  Going to debug tomorrow.

I also moved all of the database helper functions for specific tables into different classes, ProjectsHelper, ServersHelper, and LayersHelper.

I modified the AddLayersList