Changed the deleteButton in the layer_list_item to an image button.
Added a delete method to the ArbiterDatabaseHelper interface and implemented the delete method for the LayersHelper.  I also added the layer's id to the Layer to make it easier to get for deleting the layer.  I also added the onClickListener to the deleteButton in the LayersDialog for deleting that layer.

I also removed the db.close() lines from the database helper classes because the SQLiteOpenHelper handles opening, caching, and closing the db.

I also added a method to prevent layers that are already in the project from being added to the AddLayersDialog.