Moved some files around and/or renamed them.  The Layer, Project, and Server classes got moved into the BaseClasses directory.  The ProjectListAdapter got moved into the ListAdapters directory.  Added a singleton, ArbiterProject, which is used for keeping track of the currently open project.  The LayersHelper uses this in it's getAll method to get all layers from the current project.  The ProjectListAdapter uses this to add "[current]" next to the currently open project in the list of project.
I also got rid of the DbHelpers singleton and changed the GlobalDatabaseHelper, and all of the TableHelpers into singletons.

Also implemented project deletion so that when a project gets deleted, so do all of the layers associated with that project.

In addition, added a "find me" button and a "sync" button to the MapActivity's layout.  Still need to implement their functionality.