Fixed the error that was occurring due to the map not being initialized before trying to add layers to it.  Got rid of the js that was attempting to do this, and instead called the loadMap method from within the MapActivitys onMessage method which gets notified when the CordovaWebViews page finishes loading.
I also added logic checking to determine whether or not the maps layers have to be reloaded.  When the MapActivity resumes, if the open project id is different from the project id when the map was loaded, the layers will get cleared from the map, and the new layers will be added.