Made com.lmn.Arbiter_Android.Map into a singleton for holding the code to call javascript mapping related calls.
Also created an interface in the LayerListAdapter for communicating with the MapActivity.  The MapActivity implements the onLayerVisibilityChanged and onLayerDeleted methods and then the LayerListAdapter can call those to update the map in the MapActivity.  I added a callback to the LayersHelper delete method to be executed after a successful delete.  In the deleteLayer button listener, I pass a callback, which can run on the UI thread, and update the map by calling the onLayerDeleted method in MapActivity.