Modified the Map and Layer Loaders, and their callbacks, to use ArrayList's of layers instead of arrays of layers to account for the default layer being added.
Also set it up so that projects created by default will have a default openstreetmap layer.  Also handled the deletion of this layer.  Split the removeLayer method in Arbiter.Map.Layers to removeLayerById and removeLayerByName to make it easy to delete both by a layers name and by the layers id.