Added a callback argument to Arbiter.Map.createMap.  Used this as an entry point to override the initial osm base layers getURL method.
Also made sure that the savedZoomLevel and savedExtent get reset onResume for both the MapActivity and the AOIActivity, that way the same variables could be used for both activities when reseting the web app, as part of the out of memory workaround.

I also added the functionality in the AOIActivity to load the web app url when the about:blank page finishes loading, and initialize an instance of MapLoaderCallbacks to complete the reset.