Got rid of a bunch of console logs, added another couple params to the FeatureTableHelpers loadFeatures callback so I could check inside the callback to see whether or not to execute the FeaturesLoader’s loadFeatures onSuccess callback..
Added a remove method to the PreferencesHelper to delete a key from the table.

Added an event to the map to notify listeners that the layers have finished loading on the map.

Created a module to handle feature selection.  When a layer gets added to or removed from the map, the layer gets added to or removed from an array which is used for layers kept track of by the open layers select feature control.  Due to the OOM workaround, I’ll need to save the selected feature before resetting if there is a selected feature.