Made several updates:
If the area of interest hasn’t been set, use the open layers layers original getURL method instead of the overridden one from the TileUtil.  Also, don’t download features for a layer if the area of interest hasn’t been set yet.

Updated the SyncHelper for caching tiles.

The cordova file api doesn’t have the option to create parent directories as necessary, so we have to explicitly create the directories for the tile sets before caching the tiles.

Made more modifications to the TileUtil to be used with this version of Arbiter.  Things are working now, but in the future we should refactor a lot of it.

Added a method to Util.java for quick access to show an alert dialog.

Moved the logic for toggling the dialog and buttons related to incomplete projects to the MapChangeHelper

Made some updates to show the sync percentage in the sync progress dialog.  This currently doesn’t include vector data syncing in it’s calculation.

Added a confirmation dialog to confirm the user wan’t to change their area of interest.